+++ engine/src/main/java/net/jqwik/engine/properties/arbitraries/randomized/UniformRandomDistribution.java
@@ -27,4 +27,9 @@ public class UniformRandomDistribution implements RandomDistribution {
 			&& max.compareTo(BigInteger.valueOf(Integer.MAX_VALUE)) <= 0;
 	}
 
+	@Override
+	public String toString() {
+		return "UniformDistribution";
+	}
+
 }